refactor: Move pack/unpack `IP_Port` from DHT into network module.
It's misplaced in DHT, since the data structures are located in network.
