+++ other/bootstrap_daemon/docker/tox-bootstrapd.sha256
@@ -1,1 +1,1 @@
-af83f07bb96eb17a7ee69f174c9960d23758c9c9314144d73e0f57fdef5d55e4  /usr/local/bin/tox-bootstrapd
+e9e6e7baafe3cfe9b212ffb4f1ea8a8b48e2dc1ee4c07c6b1b0b04893b3b6464  /usr/local/bin/tox-bootstrapd
